crossystem: Add BINF3_LEGACY for systems boot in legacy mode.
The BINF3 (VBT7) reflects the firmware type, and we need that to
describe which type of firmware was booted.

The 'legacy' did not have its own value definition, but without that we
can't make sure if the system is running a non-chrome firmware or simply
entered legacy boot path. CL:1054307 introduced a new value (0x4) for
legacy type and we should handle it in crossystem mainfw_type command.

BUG=b:79130310
TEST=emerge-eve coreboot depthcharge chromeos-bootimage;
     Boot in legacy mode and see crossystem reporting 'legacy' for
     mainfw_type.

Change-Id: I4a1165e547e70c634d45054f56d1357ae5af2a83
Reviewed-on: https://chromium-review.googlesource.com/1068556
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>